+++ test/frontend/__fixtures__/event_price_categories.ts
@@ -1,0 +1,8 @@
+import { EventPriceCategory } from 'models/event';
+
+const categories: Array<EventPriceCategory> = [
+  { id: 1, name: 'Students' },
+  { id: 2, name: 'Partners' }
+];
+
+export default categories;