+++ test/integration/open_api/subscriptions_test.rb
@@ -9,12 +9,15 @@ class OpenApi::SubscriptionsTest < ActionDispatch::IntegrationTest
     @token = OpenAPI::Client.find_by(name: 'minitest').token
   end
 
-  test 'list all subscriptions' do
+  test 'list subscriptions' do
     get '/open_api/v1/subscriptions', headers: open_api_headers(@token)
     assert_response :success
+    assert_equal Mime[:json], response.content_type
+
+    assert_not_empty json_response(response.body)[:subscriptions]
   end
 
-  test 'list all subscriptions with pagination' do
+  test 'list subscriptions with pagination' do
     get '/open_api/v1/subscriptions?page=1&per_page=5', headers: open_api_headers(@token)
     assert_response :success
   end